Reorganize test support classes into testsupport package
Rename and restructure the test support infrastructure from
`core.testing.model` to `core.testsupport` with a clearer layout:

- testsupport/
  - MockedModelTestSupport.java (main entry point)
  - mocking/ (low-level mock factories)
  - builder/ (fluent builders for test fixtures)
  - tests/ (tests for the support infrastructure)

This makes the purpose of these classes clearer - they are test
support utilities, not tests themselves.
