+++ core/src/test/java/gov/nist/secauto/metaschema/core/testsupport/builder/FlagBuilder.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: CC0-1.0
  */
 
-package gov.nist.secauto.metaschema.core.testing.model;
+package gov.nist.secauto.metaschema.core.testsupport.builder;
 
 import static org.mockito.Mockito.doReturn;
